requirements.txt
streamlit
langchain
langchain-openai
langchain-google-genai
pypdf
pandas
openpyxl
python-dotenv
